LRQA-19505 Refactoring - changed package name
