+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/load/balance/LoadBalanceUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.jenkins.load.balancer;
+package com.liferay.jenkins.load.balance;
 
 import com.liferay.jenkins.results.parser.JenkinsResultsParserUtil;
